+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Master
 
 - Updated iOS minimum deployment target to 8.0
+- Moved `Core` pod subspec to `Foundation`
 - Removed AFNetworking dependency
 - Removed `network` key from `-[UIApplication analyticsInfo]`
 - Removed NSString methods `-contains:`, `-containsCaseInsensitive:`, `-contains:options:`