+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Removed `ESScreenSizeString(CGSize size)` function
 - Removed `screen_size` from `-[UIApplication analyticsInfo]`, added `screen_width` and `screen_height`
 - Removed `Screen` from `-[UIApplication userAgentForHTTPRequest]`
+- Removed `ESUUIDString()`
 - Generate Xcode projects through XcodeGen
 
 ## 3.18.12 (2019-11-10)